Move reporting into its own crate
