+++ Cargo.toml
@@ -14,6 +14,7 @@ members = [
     "compiler/constrain",
     "compiler/unify",
     "compiler/solve",
+    "compiler/reporting",
     "vendor/ena",
     "vendor/pathfinding"
 ]
