+++ compiler/reporting/src/lib.rs
@@ -11,10 +11,4 @@
 // re-enable this when working on performance optimizations than have it block PRs.
 #![allow(clippy::large_enum_variant)]
 
-pub mod string;
-
-pub mod crane;
-pub mod fmt;
-pub mod llvm;
-pub mod load;
-pub mod mono;
+pub mod report;
